Externalize appActions & change imports
